requirements.txt
Pillow
rembg
diffusers
torch --index-url https://download.pytorch.org/whl/cu118
cuda-python
tokenizer
transformers
requests
onnxruntime
accelerate
flask